feat: migrate CRD types from *int to *int64
Migrate all CRD field types from *int to *int64 to align with GitLab
client-go v1.0 API which uses int64 for all ID fields.

This is NOT a breaking change for users as YAML/JSON does not
distinguish between int and int64 - both serialize as numbers.

Changes:
- Update all API type definitions in apis/namespaced/
- Update all API type definitions in apis/cluster/
- Update helper functions: IsIntEqualToIntPtr -> IsInt64EqualToInt64Ptr
- Update client code to remove unnecessary int() casts
- Update controller code to work with int64 types directly
- Regenerate deepcopy code for all API packages
- Regenerate cluster-scoped files from namespaced templates

Benefits:
- Eliminates hundreds of manual type conversions
- Direct compatibility with GitLab API types
- More maintainable codebase
- Prevents potential overflow issues with large IDs

Files affected:
- All *_types.go files in apis/namespaced/ and apis/cluster/
- Client implementations in pkg/namespaced/clients/
- Controller implementations in pkg/namespaced/controller/
- Generated cluster-scoped equivalents

Related to GitLab client-go v1.0 migration.

+++ apis/cluster/groups/v1alpha1/zz_badges_types.go
@@ -29,7 +29,7 @@ type BadgeParameters struct {
 	// GroupID is the ID of the group to create the deploy token in.
 	// +optional
 	// +immutable
-	GroupID *int `json:"groupId,omitempty"`
+	GroupID *int64 `json:"groupId,omitempty"`
 
 	// GroupIDRef is a reference to a group to retrieve its groupId
 	// +optional
@@ -45,7 +45,7 @@ type BadgeParameters struct {
 	// If the badge with this ID does not exist, resource creation will fail.
 	// +optional
 	// +immutable
-	ID *int `json:"id,omitempty"`
+	ID *int64 `json:"id,omitempty"`
 	// LinkURL is the onclick redirect URL of the badge.
 	// Supports gitlab format templating using variables like %{project_name}
 	// +required
