+++ apis/cluster/projects/v1alpha1/zz_badges_types.go
@@ -29,7 +29,7 @@ type BadgeParameters struct {
 	// ProjectID is the ID of the project to create the deploy token in.
 	// +optional
 	// +immutable
-	ProjectID *int `json:"projectId,omitempty"`
+	ProjectID *int64 `json:"projectId,omitempty"`
 
 	// ProjectIDRef is a reference to a project to retrieve its projectId
 	// +optional
@@ -45,7 +45,7 @@ type BadgeParameters struct {
 	// If the badge with this ID does not exist, resource creation will fail.
 	// +optional
 	// +immutable
-	ID *int `json:"id,omitempty"`
+	ID *int64 `json:"id,omitempty"`
 	// LinkURL is the onclick redirect URL of the badge.
 	// Supports gitlab format templating using variables like %{project_name}
 	// +required
