+++ apis/namespaced/groups/v1alpha1/accesstoken_types.go
@@ -29,7 +29,7 @@ type AccessTokenParameters struct {
 	// GroupID is the ID of the group to create the deploy token in.
 	// +optional
 	// +immutable
-	GroupID *int `json:"groupId,omitempty"`
+	GroupID *int64 `json:"groupId,omitempty"`
 
 	// GroupIDRef is a reference to a group to retrieve its groupId
 	// +optional
@@ -69,7 +69,7 @@ type AccessTokenParameters struct {
 // GitLab API docs:
 // https://docs.gitlab.com/ee/api/group_access_tokens.html
 type AccessTokenObservation struct {
-	TokenID *int `json:"id,omitempty"`
+	TokenID *int64 `json:"id,omitempty"`
 }
 
 // A AccessTokenSpec defines the desired state of a Gitlab group.
